Move get_source_klass and get_source_klasses
