+++ app/connectors_service/connectors/service_cli.py
@@ -22,7 +22,7 @@ import click
 from click import ClickException, UsageError
 from connectors_sdk.content_extraction import ContentExtraction
 from connectors_sdk.logger import logger, set_logger
-from connectors_sdk.source import get_source_klass, get_source_klasses
+from connectors.utils import get_source_klass, get_source_klasses
 
 from connectors import __version__
 from connectors.build_info import __build_info__
