+++ app/connectors_service/connectors/services/job_scheduling.py
@@ -14,7 +14,7 @@ Event loop
 import functools
 from datetime import datetime, timezone
 
-from connectors_sdk.source import get_source_klass
+from connectors.utils import get_source_klass
 
 from connectors.es.client import License, with_concurrency_control
 from connectors.es.index import DocumentNotFoundError
